backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
neo4j==5.15.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# API & Data Validation
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Async Tasks
celery[redis]==5.3.4
flower==2.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
networkx==3.2.1

# Machine Learning
scikit-learn==1.3.2
torch==2.1.2
spacy==3.7.2

# Report Generation
jinja2==3.1.2
weasyprint==60.1
reportlab==4.0.7
python-docx==1.1.0

# Blockchain APIs
web3==6.11.3
eth-utils==2.3.1
eth-typing==3.5.2

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2
